Simple updates per feedback
